+++ src/action/filter/Filter.java
@@ -4,5 +4,5 @@ package action.filter;
 import action.Action;
 
 public abstract class Filter implements Action {
-    protected float ratio;
+    protected double ratio;
 }
